Exit translation on async eager return in unlikely profiling branches
Summary:
If the profiling information suggests that we almost always call a particular
function (>99.9% probability), we exit the translation after the generic call
in the unlikely profiling branch.

But we don't do the same with async eager return, which means the unknown
result of the unlikely branch will pessimize the return value from the known
(potentially inlined) function on the hot path.

Let's move the responsibility of exiting translation to the callUnknown() and
implement the same logic for async eager return.

Reviewed By: oulgen

Differential Revision: D15883382

fbshipit-source-id: 92e292f5fb559bc9218c34a2ce4eb769fbbb06b7