Add --extract-dependencies mode to hh_client with a trivial implementation
Summary: We would like to automate extracting dependencies of a given function / method to easily create self-contained test cases from examples in large codebases and, ultimately, facilitate typechecker debugging. This commit introduces an --extract-dependencies mode, which takes a fully qualified function name and prints (in the current version) the function code. This is to be extended with actual extraction of dependencies.

Reviewed By: manzyuk

Differential Revision: D15879513

fbshipit-source-id: 5558b1b168cc1370fb133cde683ae1a3d0e2e26e