Reject unset by asserting subtypes
Summary:
Replace the use of `is_sub_type_LEGACY_DEPRECATED` when checking validity of `unset`. Instead, use a subtype assertion against a union, wrapped in a try-catch to produce a good error.

Where possible, we should avoid `is_sub_type` variants. If a particular subtype relation is *required*, we should be using subtype assertions, not predicates.

Reviewed By: manzyuk

Differential Revision: D15897679

fbshipit-source-id: 5acf661e07009df1fe5d357cf89b7819bd33f045