Clean up two uses if TUtils.is_option
Summary: Replace two uses of the somewhat fragile `is_option` function. One was just to optimize the case of "nullable of nullable", which is better handled by using `TUtils.union` with the null type. The other (testing that something *isn't* nullable) is better handled by asserting that a type is a subtype of `nonnull`.

Reviewed By: manzyuk

Differential Revision: D15897876

fbshipit-source-id: d137d01f3f02bb46ebd95fa012a8ec85f2e391ad