refactor slab manager to use pages in the local arena
Summary:
Let slab manager manage a pool of 2M slabs on the same address range (and huge page) of with the local arena.  The slabs are not part of the arena, to avoid overhead of jemalloc.  It also allows us to put slabs on part of a 1GB page, and use the remaining part in a jemalloc arena.
Slab manager no longer works if not using jemalloc.

Reviewed By: alexeyt

Differential Revision: D15850238

fbshipit-source-id: a522438b6850b9f301574682e98832a56294483e