Refactor Config_file + print overrides
Summary:
I need to be able to override values in hh_server's local config and also to be able to confirm that the overrides were applied. Because stray keys don't result in failures (which is OK), I at least would like to see the before and after state of the config.

A few things here:
* Refactoring to unify some common processing steps.
* Adding printing for overrides
* Adding an LWT implementation for clients - I originally added it to the same module, `Config_file`, but then the exciting EINTR error reappeared [see this diff and Jake's comment at the end](https://our.intern.facebook.com/intern/diff/D14099177/?transaction_id=780727162283493&dest_fbid=2041359402649022).

Differential Revision: D15800893

fbshipit-source-id: 0ad90726db8336689680822267ae50666e4db0fe