Redo how we do incremental naming table saved saved states.
Summary: The existing method of making incremental saved states for naming involves too many table scans.  In tests it completed in a reasonable amount of time, but in practice it was taking over an hour (sixty times longer than just generating one from scratch!).  With this change, we just persist the local delta in SQLite.

Reviewed By: 2BitSalute

Differential Revision: D15867348

fbshipit-source-id: 0da9aa0e1c3daf46d1efadc6af21596fdcaf3a9a