No longer check for naming table existence when mounting DB.
Summary: In some incremental saved state workflows, we load in an empty table before copying over every entry.  This check was originally added to make debugging easier, but now we can just depend on SQLite failing to read if we try to read from a database without the proper tables.

Reviewed By: 2BitSalute

Differential Revision: D15904279

fbshipit-source-id: 292735c721ed5dd251b496e3fb35dfa0562aafbd