Fix duplicate synthesized elements in MRO
Summary:
There is a copy-paste bug here. When we encounter an element that has already been emitted as part of the current linearization, we want to skip it--there is no reason to emit it twice.

While emitting a linearization lazily, we do indeed skip it (by using `Skip` rather than `Yield` here). However, we always add it to the accumulator containing the entire (reversed) linearization, which we store to shared memory upon completion. This means that we write duplicate MRO elements to shared memory, wasting space and time.

This change stops adding skipped elements to the accumulator.

Reviewed By: Wilfred

Differential Revision: D15891914

fbshipit-source-id: a56e4af7ec17b72c5b4d04cf836ac143c96ffdd8