Forbid trait reuse
Summary:
Reimplement the detection of trait reuse for shallow_class_decl.

Because Decl_linearize already abstracts over the reuse of any ancestor (by not emitting an mro_element for any ancestors previously emitted in the linearization), we must emit a marker mro_element when we encounter a reused trait if we would like to report the error without 1) reporting errors during linearization or 2) reimplementing the ancestor traversal where we would like to report the error.

Differential Revision: D15167015

fbshipit-source-id: 89fb7707af20c81e04a30bb13bc2b1cf88d8068d