Dedup synthesized mro_elements earlier
Summary:
Currently, we accumulate all synthesized mro_elements in a separate list (here, `synths`) as we emit non-synthesized elements, then emit synthesized mro_elements once we know all non-synthesized elements have been emitted. We dedup synthesized elements as we do with all other MRO elements--we iterate over the accumulated linearization-so-far (`acc`) looking for an element identical to the one we are about to emit.

This is a little wasteful. Linearizations are mostly composed of non-synthesized elements. There is no need to compare a synthesized element to non-synthesized ones, since they will never be equal. Since we already accumulate synthesized MRO elements in a separate list as we emit the non-synthesized ones, we can iterate over that list to dedup instead.

Differential Revision: D15891913

fbshipit-source-id: cddee25cf9463000ec015b989c93874efcf65224