Check sealed whitelist
Summary:
When shallow_class_decl is enabled, Decl_heap.Classes will be empty, so existing uses of it (and its value type, decl_class_type) will be no-ops. This causes us to miss errors which are currently detected only by examining decl_class_type.

Instead, we should use Decl_provider.Class.t, which abstracts over whether shallow_class_decl is enabled.

The only remaining use of decl_class_type in Typing is when we check the sealed whitelist on parent types (to verify that the child class is permitted to extend it). Typing_classes_heap does not yet provide an accessor for the sealed whitelist, so we add one.

Reviewed By: vassilmladenov

Differential Revision: D15167014

fbshipit-source-id: 476ff816eb136c3b42aa6de7fc0f7bb14460ecb9