Fix newline-counting bug in formatting-suppression logic
Summary: In D14786882, we attempted to ensure that nodes with 3 or more newlines within them are formatted even if they are preceded by an error-suppression comment. However, this `count = n` assertion only makes sense if we check every time we increment `count`, but we don't. We have a couple `List.fold`s below where we don't check after each increment. This causes us to incorrectly refuse to format some large nodes preceded by a FIXME.

Reviewed By: periodic1236, pittsw

Differential Revision: D15910236

fbshipit-source-id: 66c4700c501a01fdca8b6675ecc081198ae2ed2a