Check source and bytecode SHA1 before evicting units from cache
Summary: If the source text hasn't changed there's no reason to recompile the bytecode. Likewise if the bytecode hasn't changed we can avoid throwing away the old unit (and all of the cached classes, functions, and translations associated with it). Also added more context to `hh\get_compiled_units()` to test for files that were recompiled but not evicted from the cache.

Reviewed By: jano

Differential Revision: D15904204

fbshipit-source-id: b42ab36bc7c8c3c67306834ad6a3e869fd96a4b1