simplify bounds containing the type variable itself
Summary:
Remove a type variable from its upper and lower bounds.
More precisely, if a type variable appears in one of its bounds under any combination of unions and intersections, it can be simplified away from the bound.
For example,
- if #1 has lower bound (#1 | A), the lower bound can be simplified to A
- if #1 has upper bound (#1 | B), the upper bound can be simplified to mixed

and dually for intersections.
See inline function documentation for how this is done.

Reviewed By: manzyuk

Differential Revision: D15851856

fbshipit-source-id: 16a36e9385dba28e2d6475382324b1a0fb864764