Generalize "pos" field in Vinstr
Summary:
Vinstr has a "pos" field which is just an unsigned int. Its used by
XLS and vasm-graph-color to store various metadata. Formalize this
concept a bit. Rename it from "pos" to "id" because it does not always
represent a position for XLS. It will soon be used for caching by
vasm-graph-color (right now its an index into the pseudo instruction
table). Some vasm utilities might mutate a Vinstr, and depending on
what "id" represents, might invalidate it. For such utilities, add a
new optional "clobber" parameter. If provided, "id" will be set to
that value if the instruction is mutated. vasm-graph-color will use
this to ensure the instruction's cache entry is invalidated if the
instruction changes.

Reviewed By: paulbiss

Differential Revision: D15863363

fbshipit-source-id: dd86e95c4cec0ab77fb4a06916639d69aa4c86ef