Type_test_hint_check not dependent on how Taccess is localized
Summary:
Currently `Type_test_hint_check` works for `Taccess` types by localizing and looking for an upper bound that is an expression dependent type, breaking the generic into its parts and then looking up the class/type const pair for the Enforceable attribute. This places a strong dependency of this code on how we localize type accesses which I want to change in subsequent diffs.

To break this dependency I update the code to localize up to the last type constant in a type access chain and looking for concrete upper bounds. I then do the class/type const look up on these upper bounds.

As a result of this change, the position on error messages actually improved to point to exactly the type constant where the Enforceable attribute would need to be added.

Reviewed By: vassilmladenov

Differential Revision: D15902357

fbshipit-source-id: 45ae0017ff6029d02c8ff64d7ae27207c3310e81