Introduce record type in HHBBC and repo-authoritative type
Summary:
The goal of this diff is to get object property, parameter and return typehints
work for records in repo-auth mode. The HHBBC record type has a string type tag
(name of the record) which seems sufficient for this purpose, but when we
introduce record inheritance, we might need a dedicated type tag for records.
Similarly, the repo-auth type currently does not have any information on record
types, which will change in future.

Reviewed By: paulbiss

Differential Revision: D14962157

fbshipit-source-id: fec93947a839e4077a4662ead4affd94f5744a91