Rename inheritable_elt to tagged_elt
Summary:
I wanted to use a struct here to give a name to the `should_chown_privates` bool, but I had trouble coming up with a name for it. `inheritable_elt` doesn't make much sense, because all class members are inheritable.

Instead, I'm suggesting `tagged_elt`, since it associates a `class_elt` with a tag indicating whether it should still be visible when the `class_elt` is private. This change also renames that tag from `should_chown_privates` to `inherit_when_private`, which I think is a bit clearer.

Reviewed By: Wilfred

Differential Revision: D14965571

fbshipit-source-id: 9869274952f4c57198b95be2e80507e62d54f59d