Remove typeconst_seen from Typing_taccess.env
Summary: As part of refactoring typing_taccess, I want to eliminate the information we track in `Typing_taccess.env`. In the case of the `typeconst_seen` field it's only use is for go-to-def for the IDE. Instead of relying on the expansion code to compute `typeconst_seen`, I use a similar technique as in {D15902357} and look for the referenced type constant by looking for concrete super types of the root. In my testing it also seems to provide better position information compared to the current implementation.

Reviewed By: vassilmladenov

Differential Revision: D15902826

fbshipit-source-id: c9f7894adb493d32d97d009bf7aa259681e3fbdd