Allow different inlining cost factors for main/cold/frozen areas
Summary:
When computing the cost of code for inlining purposes, it doesn't make sense to
count the hot code the same as the code that almost never executes. Let's have
a cost multiplier for each of the vasm areas.

This diff doesn't change the behavior yet, as all multipliers are currently set
to the same value. It allows experimentation via config updates.

Reviewed By: ricklavoie

Differential Revision: D15915891

fbshipit-source-id: 95d2922523c29bada1ad23f2bbf4190053b86c99