Reduce TreadHashMap load factor
Summary:
When I tried to land the Shapes::idx array profile optimizations previously, it caused a performance regression because of an increased number of entries in the s_inlineStacks map (used for backtracing). Note that the Shapes::idx rarely appears in error traces, but adding to this map slows down other backtracing.

I've tried a few fixes for this issue, and the simplest one is just to reduce the load factor on this hash table. Doing so makes sense because the table is small and is read much more often than it's written to. Note that although the Shapes::idx change has the worst effect of any change on lookups on this map, other inlining changes also hurt backtrace performance, as we saw with the most recent release "var_export".

Reviewed By: mxw

Differential Revision: D15824709

fbshipit-source-id: 30bcbc7e8e88c814eac3285fbc525b787824e89b