Make Shapes::idx a native builtin
Summary:
This diff uses FCallBuiltin to inline calls to Shapes::idx in all translations. Doing so is important for further optimization, because the TargetProfile PGO data structures are bound to a particular bytecode index; currently, all Shapes::idx profiling data gets merged together, even though each callsite sees a different shape of array.

I previously landed a version of this diff that looked good in smaller benchmarks but was a regression in prod. I addressed this regression by decreasing the load factor on TreadHashMap in a previous diff, to optimize for reads (this map is written to when we JIT new code but read whenever we throw an exception). I've also split up the diff into three smaller pieces to make it easier to land.

There was another flaw in the original diff: I missed some Hack array compatibility notices. I have two fixes for that problem. First off, I added a number of tests that raise notices. I was also able to rebase on top of some of Jan's work to harden type errors and then reuse the normal builtin type-checking logic, making me much more confident in continued correctness. Finally, I also used the new "native" arguments rather than Variant to take inputs in Shapes::idx, since the register passing fix is also in.

Reviewed By: mxw

Differential Revision: D15695962

fbshipit-source-id: 4f38753c3d88e578138e452b0096eff26b49014a