Subtyping of nullable <: union
Summary:
We are missing a case in subtyping: `?t <: t1 | ... | tn`. If one of the `ti's` is `null` then this is rejected.

I don't have a Hack repro for an issue, but can observe the problem by constructing internally a type of the form `Tunion [Tnull; t1; t2]`, for example, in D15918173.

Reviewed By: manzyuk

Differential Revision: D15921966

fbshipit-source-id: 1216f84ace139ad2fbe6c2bd06337552f9f4a835