Better error messages for disjunctions: propagate the error back to the constraint from which the disjunction was generated
Summary:
Error messages arising from disjunctions of subtyping propositions can be confusing.  For example, the subtyping constraint `int <: (dynamic | string)` simplifies to a disjunction `(int <: dynamic) | (int <: string)` but because the first constraint is patently unsatisfiable the smart constructor `Typing_logic.disj` for disjunctions discards it and we end up with `int <: string`, which is also unsatisfiable, and we show the error "Expected string but got int" to the user, even though it would be more precise to say e.g. "Expected dynamic or string but got int" (or "Expected ~string but got int" if `(dynamic | string)` arose as elaboration of the like type `~string`).

The way I'm proposing to improve things is by storing in the disjunction constructor (a suspension of) an error message that should be generated if none of the branches of the disjunction can be satisfied.  (Incidentally, this makes `Typing_logic.Unsat f` equivalent to `Typing_logic.Disj (f, [])`, and I'm planning to remove the former in a subsequent diff.)

Reviewed By: andrewjkennedy

Differential Revision: D15897521

fbshipit-source-id: 27239c04e580f4d73b8ea9e85f13be608de60c31