Unified symbol-to-docblock server command
Summary:
The "completion/resolve" command required two separate roundtrips to the server to calculate a docblock.  That's a little bit wasteful.  Let's do it with one roundtrip instead.

Previously, we used LOCATE_SYMBOL followed by DOCBLOCK_AT.  Now we'll just use a single command, DOCBLOCK_FOR_SYMBOL.

While working on this I also noticed that the documentation for the "data" element in LSP completion / resolve says that the value is only intended to be used as output from textDocument/completion for processing in completion/Resolve.  Therefore, we no longer need to do fixup on that value when it comes out of completion/Resolve.

Reviewed By: pittsw, arxanas

Differential Revision: D15910822

fbshipit-source-id: 497f5ae866efbd30f51d29beefc6dbe75ea29f12