When passed a function with --extract-standalone, extract and print its dependencies.
Summary: We would like to automate extracting dependencies of a given function / method to easily create self-contained test cases from examples in large codebases and, ultimately, facilitate typechecker debugging. This commit finds dependencies of the function passed with --extract-standalone flag. In the current version those dependencies are just printed, and in future we will produce minimal Hack code for them.

Reviewed By: 2BitSalute

Differential Revision: D15920923

fbshipit-source-id: fac16d492dc7e6fa80da706b54cc646f0885a94c