Remove copyargs vasm simplification rule
Summary:
vasm-simplify will lower a copyargs into a set of single copies if the
sources and destinations are all disjoint. This seems to buy very
little and actually over constrains (the copyargs are semantically
done in parallel while the copies are serial). Leaving copyargs as is
gives more freedom to the register allocator, so don't perform this
simplification.

Reviewed By: ottoni

Differential Revision: D15863615

fbshipit-source-id: 0aa01ccc32095f94389b21c5a5d137c3b336dbb5