Invalidate files via hh_server [with options]
Summary:
Adds a step in the typechecking phase of hh_server that makes POST requests to HHVM's default admin port: localhost:8088 to invalidate units.

This is to be used to determine whether people notice a significant difference in their user experience with invalidating files based on type dependencies.

This is enabled by two flags in `.hhconfig`: `tico_invalidate_files` and `tico_invalidate_smart`. With `tico_invalidate_files` enabled, after typechecking, all of the changed files and their dependents (based on file level dependencies) will be invalidated via a post request. With both `tico_invalidate_files` and `tico_invalidate_smart` enabled, only files that hh_server has typechecked will be invalidated. Since hh_server uses finer grain information to make its determination about which files to recheck, there will likely be fewer invalidated files per change.

Since there is a 100 MiB limit on file dependencies, we batch files in batches of 500,000 - which can still exceed the limit if the paths are too long, but for the moment, we see about it clocking in around 70MiB max. For each batch, we use a new socket because there were issues reusing the same socket to make all of the requests.

Reviewed By: Wilfred

Differential Revision: D15818900

fbshipit-source-id: 206173bacc9307a57443824b59ab564df0a47cec