Implement `Identify_symbol` using an AST as input
Summary: Make it possible to pass an AST into `Identify_symbol` instead of a file name. This is a proof-of-concept for the serverless IDE, which expects to pass in the contents of the IDE buffers where necessary, rather than relying on the file heap/parse heap/decl heap to have the right contents.

Reviewed By: 2BitSalute

Differential Revision: D15059410

fbshipit-source-id: 3013cdfb7bde2f987bc2ccd04203428858faf5fb