Create `Memory_bounded_lru_cache` as an alternative provider backend
Summary:
For the serverless IDE, we want to be able to control memory usage in a fine-grained way. This commit creates a cache that evicts entries in LRU order, such that we remain below a pre-specified memory limit.

This cache allows us to store many types of entries in a heterogeneous way, so that we can cap overall memory usage, instead of per-heap.

Reviewed By: pittsw

Differential Revision: D15169751

fbshipit-source-id: 73554bfd3f9217c5db0510f14c90feb3edfca2fd