Support local-memory caching for `Decl_provider`
Summary:
Support caching decls in local memory for the serverless IDE.

We expect this to be the only thing that needs to be cached. File contents, ASTs, etc. can be read on-demand to reduce memory overhead, and these operations don't happen often. (We probably will never need the AST for a file after getting its decls except for some IDE requests -- in which case, we will only need the AST for one file rather than a large batch of files).

This duplicates much of the logic in `Typing_lazy_heap`, except that it checks our local memory caches instead of the shared memory caches.

Reviewed By: 2BitSalute

Differential Revision: D15404424

fbshipit-source-id: 71b3b785a0847e9b1f282765f5171fd317cad74b