Add trivial constant optimization to vasm-graph-color
Summary:
vasm-graph-color does a somewhat complicated analysis to find
the best point to materialize constants. However this is unnecessary
for a common case. That is, if the constant is used exactly one place,
and that place isn't within a loop. Such a use is called "trivial" and
the best place to materialize it is immediately before its use. Add
logic to the constant placement phase to identify such trivial
constants and avoid the expensive dataflow for them. This speeds up
the pass significantly.

While I'm at it, perform one other optimization. The register
allocator cannot do anything with reserved registers. So, if you have
a copy instruction with a constant as the source, and a reserved
register as the dest, we'll materialize the constant into an
unreserved register, and then copy it into the reserved register. This
takes two instructions and uses an extra register. Recognize this case
and materialize the constant directly into the reserved register.

Reviewed By: mofarrell

Differential Revision: D15864046

fbshipit-source-id: 10b11f678748b98186fb645d8eb99b2c07346159