Fix HHBBC DCE of NewKeysetArray and family
Summary:
HHBBC was DCEing NewKeysetArray and other array-like creation opcodes,
even if they might raise an error. Fix this by only performing the
elimination if the opcode won't throw.

We weren't marking all of the opcodes which might not throw, so do
that to prevent regressions. NewKeysetArray will only throw if given a
bad key, and ColFromArray can raise if given an improper array type.

As part of this, I noticed that keyset[null] was not being marked as
unreachable, even though something like keyset[false] was (both will
throw). The reason is because maybe_class_func_key returns a type for
TNull, even though TNull isn't a class or func pointer. Fix that to
handle TNull properly.

Add EntryPoint function to invalid-scalar.php test case, which
demonstrates the problem (it wasn't an issue before because being in
the pseudomain pessimized DCE).

Reviewed By: markw65

Differential Revision: D15958428

fbshipit-source-id: 45b5e204f2331ad820f2ca1669b557100340e0f1