Fix creation of expression dependent types from nullable types
Summary: If `C` is a class and `$c : C` is passed as an argument to a function / method, the type of `$c` is replaced with an expression dependent type `<expr#n> as C`.  This construction is distributed over unions, so if `$c` has type `C | D`, it is replaced with `<expr#n> as C | <expr#n> as D` (which may later be simplified by functions from `Typing_union.union` to `<expr#n> as (C | D)`) but not over options (nullable types): if `$c` has type `?C`, it is replaced with `<expr#n> as ?C`, which in particular makes `null` incompatible with the type of `$c`.  (See a new test case below.)  Options are really unions (and eventually we will replace the former with the latter), so they should at least be handled in compatible ways.

Reviewed By: dlreeves

Differential Revision: D15964813

fbshipit-source-id: fa40c823b3c0281a5edef1975f11f4a01d4f279c