Improve error message on misplaced async modifiers
Summary:
For syntax errors due to a misplaced `async` modifier, inform the user what they should do.

Also improve the error message so it's specific to the current code (interface methods versus abstract methods).

Remove the corresponding NastCheck for async modifiers, as it was dead code (we stop at the syntax error).

Differential Revision: D15920020

fbshipit-source-id: 3aa80a68ade2dafafae3d65f22403c73f372698b