Fix list destructuring of unions
Summary:
List destructuring of unions was implemented by creating a tuple of fresh types as long as the list, and then subtyping with the type being destructured. This works fine for unions of tuples, but does not work for all the cases that a `list` expression can destructured.

Instead, this runs the `assign` function for each type in the union, and then merges the list of tuple types resolved into a tuple of unions, which is then assigned again to the original list.

Reviewed By: manzyuk

Differential Revision: D15956889

fbshipit-source-id: 6c381c9965f4af0fcda882465322d3cf5bb4e647