`Deferred_decl` module: Storage for dependencies and dependents
Summary:
Today, when we type check file `A.php`, we declare dependencies as we encounter them. This leads to bad type checking times on files that depend on a lot of classes. This is relevant in the case of generated files, for example.

The idea of this diff stack is to defer declaring dependencies and discard type errors for `A.php` if its dependencies weren't already available.

=====

The `Deferred_decl` module provides global storage for files whose declaration should be deferred.

Differential Revision: D15862879

fbshipit-source-id: 9204d16c3887e879788b3bd500b646cdb1d8ebea