Type checker options: Defer class declaration
Summary:
Today, when we type check file `A.php`, we declare dependencies as we encounter them. This leads to bad type checking times on files that depend on a lot of classes. This is relevant in the case of generated files, for example.

The idea of this diff stack is to defer declaring dependencies and discard type errors for `A.php` if its dependencies weren't already available.

=====

Here, I add a type checker option to allow deferring declaring classes during type checking.

Differential Revision: D15869525

fbshipit-source-id: 3a426c8777c08d889154fcb8f9ada27739b4ee4d