Pessimize unenforceable types at function return boundaries
Summary:
This steps ahead of localization to alter the given decl type based on how much we trust its components. We call this process pessimization of types. The high level summary is that types are wrapped in like from the first enforceable type downwards. So, a `shape('a' => int)` becomes `~shape('a' => ~int)`. Consider the following example:

```
function f(): shape('a' => int) { ... }
function g(): shape('a' => int) {
  $s = f(); // $s: shape('a' => int)
  $a = $s['a']; // $a : int
  $new_s = shape('a' => $a); // shape('a' => int);
  return $new_s;
}
```
Now, we know shapes aren't enforced by HHVM, so we decide to make them like on the outside, but not the inside.
```
function f(): shape('a' => int) { ... }
function g(): shape('a' => int) {
  $s = f(); // ~shape('a' => int)
  $a = $s['a']; // $a : ~int because it came from the ~shape
  $new_s = shape('a' => $a); // shape('a' => ~int);
  return $new_s; // invalid return type: shape('a' => ~int) </: ~shape('a' => int)
}
```
Because the typechecker has historically trusted unenforceable types, if we now choose not to trust a compound type, we **cannot** trust its component types. Another note is that all erased generic type parameters are now implicitly `super dynamic`, and their `as` constraints are pessimized.

```
function f<T as arraykey>(vec<T> $v): void {
  // old $v: vec<T>, T <: arraykey
  // new $v: vec<T>, dynamic <: T <: ~arraykey
}
```

Not all localization functions are implemented yet as some are used in places where HHVM takes the exact type, such as `as` expressions. Notable pieces of pessimization not implemented by this diff:
- where constraints
- explicit type arguments for functions

Reviewed By: Matt-Schellhas

Differential Revision: D15436137

fbshipit-source-id: a17e8a7901ee1a64bb89298b678010c1a0650cf7