HHVM Debugger: Fix debuggers not handling async functions on the call stack
Summary:
The stack frame methods in ExecutionContext are not currently async-aware. As a result, when stopped at a breakpoint in an async function, selecting a call frame higher up in the stack above an await yields no variables in scope (or occasionally a $awaitable AwaitHandle that is opaque and the user can't look into). Additionally, attempting to evaluate any expression in that context fails with a cryptic error.

This diff has ExecutionContext use the logic in backtrace.cpp to properly walk up a stack containing async functions.

Reviewed By: paulbiss

Differential Revision: D15890730

fbshipit-source-id: 1fccc83a9b318ec284035da7dcd7c6ed7a167f81