Always serialize KindOfClsMeth like varray
Summary:
KindOfClsMeth should be serialized like varrays

What about EvalHackArrDVArrs? This is "the switch" for the Hack Array Migration that will create unmarked vecs instead of varrays. For serialization and similar code points though, we'll want to treat unmarked vecs like with did varrays and marked vecs like with did vecs. VariantController will need to be changed further to account for this, but if we want to treat KindOfClsMeth like varray, then it should fb_serialize like one.

Reviewed By: yujunglo

Differential Revision: D15909209

fbshipit-source-id: 3ae3dfcf5061f967778dad553c3a9df2bee8400d