Build ClassInfos bottom up
Summary:
The way the index builds the ClassInfos is currently highly
recursive. It iterates every class, then recursively builds ClassInfos
for each parent/interface/trait, and then recurses again for each
parent/interface/trait in order to pin down which instance of each one
to use - even though in almost all cases, classes are unique (this
last step was O(n^2) in the total number of parents + interfaces +
traits, *per instantiated ClassInfo*). In addition, there is a check for
cycles which is too strict - resulting either in failing to instantiate valid
classes, or asserting, depending on the structure, and the exact order
the classes are processed in (the new test case would usually crash).

Instead, build a dependency graph, and a ready list; avoid recursing
unless there's more than one instance of a dependent class. If any
classes remain after processing the ready list, try to break pseudo
cycles (eg A extends B, and B' extends A' - we first process B and A',
and then it *looks* like neither A nor B' is ready, since they each
have the other as an unprocessed dependency; but all the *valid*
instantiations *are* ready to be processed).

Structuring it this way should also make it possible to parallelize
this step, which I'll try in a later diff.

Reviewed By: ricklavoie

Differential Revision: D15985763

fbshipit-source-id: 4dd64b87f2503c99581be2cd60f8ae5bcc4fc926