Handle multiple upper bounds properly in Typing_taccess
Summary:
If a generic has multiple constraints and only one of them has a type constant, then we would report an error. For instance:

```
interface I1 {
  abstract const type T;
}

interface I2 {}

function test<Troot as I1 as I2, Tc>(
  Troot $x,
  Tc $y,
): void where Tc = Troot::T {}
```

When expanding `Troot::T` we would attempt to expand `I1::T` and `I2::T` since `I1` and `I2` are upper bounds on `Troot`. Before we would report an error because `I2::T` is not valid type, but `I1::T` is which should be enough to allow `Troot::T` to be valid.

I implemented this by utilizing an exception `NoTypeConst` which takes a unit function that when applied will report an error. When considering upper bounds on a generic we catch this exception and build a list of all the errors and types. If there was at least one successful type constant found in one of the upper bounds we ignore the errors, otherwise we pick one of the errors to report.

Note:
I discovered this while doing some clean ups I want to do with how the `this` type is represented.

Reviewed By: CatherineGasnier

Differential Revision: D15574343

fbshipit-source-id: 16c206fb9944a0427590dd65b353988178ad4f2b