Avoid overflow using elastic stack (exponentially increasing)
Summary:
Design & implement a mechanism to catch the deep recursion early
via `StackLimit`'s `set` (initially) and `check_exceeded` (repeatedly)
that is agnostic to threading but still lightweight (lock-free).

Run the parser by allotting stack space slightly below the actual limit,
then repeatedly increase the stack space (using custom threads) until
it's sufficient to accommodate the deeply recursive parse on degenerate
inputs.  The running time using this algorithm (exponential backoff)
is guaranteed to stay linear (if the rest of the parser is linear);
in fact, since we're doubling the stack the time is: `t + 2*t + 4*t + ...`
where the total parse time with unbounded stack is `T=k*t`, which is:

   (k/2^n + ... + k/4 + k/2 + k) * t < 2 * T (twice as slow)

(In reality, parsing is superlinear so it will be faster than that.)

Note that for most inputs there is no overhead introduced, since
the parsing succeeds on the first attempt -- within the main thread -- and
the checks are pretty infrequent and only update 1-2 thread-local integers.

Reviewed By: shiqicao

Differential Revision: D15947195

fbshipit-source-id: ca4f270f6806be711841d0ff2cab4b03da8bc641