Rename symbol index environment object
Summary:
We'd also like to enable a user to switch back and forth between symbol index providers, or re-initialize a symbol index after a large rebase, or after errors are detected.

To do this, we'd like there to be a single "symbol index context" object.  This object should carry all information and static objects used by all of the symbol index providers.

To get there, I'll have to do a few steps of refactoring.  I'll start by re-using the "local symbol index" environment, and gradually expanding it to handle all the other objects.

This diff simply renames the local symbol index "env" object to a new type: `si_env`.

Reviewed By: arxanas

Differential Revision: D15974273

fbshipit-source-id: cbf8e0764dd4153bbfec66a0b6f1b8343a605128